All sources need for Inspector are compiled. not tested yet.
* Some files moved, especially
* Inspc/Srv from Inspc, that is the service or server.
* Inspc/TargetProxy from make/InspcTargetProxy: It is not a special application, the sources are always offered to use.
  For example for a main board, which has a small board as slave.
* InspcTargetSimple is left as own directory, because a simple target need only that.
* Ipc/Ipc2c from own Ipc2c-folder: It is always a sub source, the characteristic: comes from Java2C
* Some adaptions, see details in files compared with pre-version. Only adaption.

+++ emC/Base/ExcThreadCxt_emC.h
@@ -253,7 +253,7 @@ extern_C ThreadContext_emC_s* getCurrent_ThreadContext_emC ();
 */
 #define STACKTRC_ENTRY(NAME) \
   MAYBE_UNUSED_emC struct ThreadContext_emC_t* _thCxt = getCurrent_ThreadContext_emC(); /*It is a variable in Stack. */\
-  int stacksize = ((MemUnit*)&_thCxt) - _thCxt->topmemAddrOfStack; \
+  int stacksize = (int)(((MemUnit*)&_thCxt) - _thCxt->topmemAddrOfStack); \
   if(stacksize <0) { stacksize = -stacksize; } \
   if(stacksize > _thCxt->stacksizeMax) { _thCxt->stacksizeMax = stacksize; } \
